--- tests/data/conf0/config.yaml
@@ -1,8 +1,0 @@
-
-addrbook:
-  site0:
-    url: https://example.com/data.json
-    user: user
-    password: secr3t
-    timeout: 60
-
